requirements.txt
Flask==1.1.2
praw==7.1.0
python-telegram-bot==13.1
gunicorn==19.9.0
Click==7.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.6
itsdangerous==1.1.0
alembic
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
psycopg2-binary
wheel
babel
